+++ src/@primitives/FeedItemCard/Category/Category.stories.js
@@ -2,9 +2,9 @@ import React from 'react';
 import { View } from 'react-native';
 import { storiesOf } from '@storybook/react-native';
 
-import MediaCard from './';
+import Category from './';
 
-storiesOf('@primitives/MediaCard/Component', module)
+storiesOf('@primitives/FeedItemCard/Category', module)
   .add('Default', () => {
     const centered = {
       justifyContent: 'center',
@@ -13,19 +13,9 @@ storiesOf('@primitives/MediaCard/Component', module)
       backgroundColor: '#f7f7f7',
     };
 
-    const cardDimensions = {
-      height: 400,
-      width: '92%',
-    };
-
     return (
       <View style={centered}>
-        <MediaCard
-          style={cardDimensions}
-          title={'Promised Land'}
-          category={'Series'}
-          image={'https://picsum.photos/600/400'}
-        />
+        <Category type={'Series'} />
       </View>
     );
   });
