+++ src/@primitives/FeedItemCard/Category/Category.tests.js
@@ -4,7 +4,7 @@ import renderer from 'react-test-renderer';
 import ThemeProvider from '@primitives/ThemeProvider';
 import Category from './';
 
-describe('the MediaCard Category component', () => {
+describe('the FeedItemCard Category component', () => {
   it('should render', () => {
     const tree = renderer.create(
       <ThemeProvider>
